+++ lib/view_component/collection.rb
@@ -7,7 +7,6 @@ module ViewComponent
     include Enumerable
     attr_reader :component
 
-    delegate :format, to: :component
     delegate :size, to: :@collection
 
     attr_accessor :__vc_original_view_context
@@ -39,6 +38,12 @@ module ViewComponent
 
     def each(&block)
       components.each(&block)
+    end
+
+    # Rails expects us to define `format` on all renderables,
+    # but we do not know the `format` of a ViewComponent until runtime.
+    def format
+      nil
     end
 
     private