+++ test/sandbox/app/controllers/integration_examples_controller.rb
@@ -72,6 +72,10 @@ class IntegrationExamplesController < ActionController::Base
     render(UnsafePostambleComponent.new)
   end
 
+  def multiple_formats_component
+    render(MultipleFormatsComponent.new)
+  end
+
   def turbo_stream
     respond_to { |format| format.turbo_stream { render TurboStreamComponent.new } }
   end